Add files via upload
This risk management calculator for making orphan works available under various Creative Commons Licences was created in 2012 in the UK, using Jisc funding. it is no longer online, but available via the WayBack Machine - here: https://web.archive.org/web/20150501144628/http://www.web2rights.com/OERIPRSupport/risk-management-calculator/ The tool needs updating to reflect GDPR, refinements to the criteria which are used to evaluate risk, and also the tool is not mobile device compliant. As the Director and inventor of the tool, I am happy to work in-gratis, to help relaunch this tool. I can be contacted at: naomi@naomikorn.com